+++ src/components/Home/Header.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
-import CommuteModal from '../CommuteModal/commuteModal';
+import CommuteModal from '../Commute/CommuteModal';
 
 import firebase, { auth } from '../../api/firebase';
 import { useAuthState } from '../../contexts/AuthContext'; // 인증 상태 가져오기
