+++ src/styles/Commute/CommuteModalSC.ts
@@ -1,4 +1,3 @@
-// styles/CommuteModal/commuteModalStyles.ts
 import styled from "styled-components";
 
 export const HeaderButton = styled.button`
